+++ src/main/java/org/folio/am/exception/ApplicationInstalledException.java
@@ -1,0 +1,8 @@
+package org.folio.am.exception;
+
+public class ApplicationInstalledException extends RuntimeException {
+
+  public ApplicationInstalledException(String message) {
+    super(message);
+  }
+}