+++ NEWS.md
@@ -16,7 +16,7 @@
 * Migrate CI from Jenkins to GitHub Actions Maven central workflow (MGRAPPS-96)
 * Upgrade module to SpringBoot4.0 and Spring7.0 (MGRAPPS-102)
 * Migrate Kafka producer integration to `folio-kafka-producer` and move topic config to `application.kafka.producer` following applications-poc-tools Kafka module split (APPPOCTOOL-85)
----
+* Add `cleanup` endpoint for unused Applications (MGRAPPS-100)
 
 ## Version `v3.0.0` (12.03.2025)
 * Add module permissions to bootstrap API (MGRAPPS-48)